requirements.txt
# requirements.txt - Centinel (versión completa enero 2026)
# Dependencias para scraping, análisis, dashboard, alertas y visualizaciones

# Core y scraping
requests>=2.31.0               # Peticiones HTTP: snapshots oficiales
playwright>=1.49.0             # Fallback scraping con navegador headless
playwright-stealth>=1.0.6      # Modo stealth para evitar detección anti-bot
apscheduler>=3.10.4            # Scheduler para healthchecks
flask-limiter>=3.7.0           # Rate limiting para el dashboard/app
httpx>=0.27.0                  # Cliente HTTP para healthchecks

# Procesamiento y análisis de datos
pandas>=2.1.0                  # Procesamiento principal: diffs, tablas por depto
numpy>=1.24.0                  # Cálculos numéricos (Benford, arrays)
scipy>=1.11.4                  # Estadística avanzada: Benford, chi-squared, tests
scikit-learn>=1.5.0            # Detección de outliers (Isolation Forest)
qrcode[pil]>=7.4.2             # Generación de QR para verificación de hashes
reportlab>=4.2.5               # Generación de PDFs institucionales

# Visualizaciones
# (removido el dashboard local en esta versión)

# Utilidades generales
python-dateutil>=2.8.2         # Manejo robusto de timestamps y fechas
PyYAML>=6.0.2                  # Configuración en YAML
web3>=6.20.0                   # Cliente Web3 para Arbitrum/Ethereum
eth-account>=0.11.0            # Firma de transacciones
pycryptodome>=3.20.0           # Keccak256 para Merkle Trees

# Testing y desarrollo
pytest==8.3.3                  # Suite de pruebas unitarias e integración

# Extras útiles (opcionales pero recomendados)
rich>=13.7.0                   # Logging bonito en consola durante desarrollo
